backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydicom
numpy
pillow
torch
torchvision
requests
